src/test/java/com/leetcode/arrays/medium/TheMaze_490Test.java
package com.leetcode.arrays.medium;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class TheMaze_490Test {

  private final TheMaze_490 solution = new TheMaze_490();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(
              new int[][] {
                {0, 0, 1, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 0, 1, 0}, {1, 1, 0, 1, 1}, {0, 0, 0, 0, 0}
              },
              new int[] {0, 4},
              new int[] {4, 4},
              true),
          arguments(
              new int[][] {
                {0, 0, 1, 0, 0}, {0, 0, 0, 0, 0}, {0, 0, 0, 1, 0}, {1, 1, 0, 1, 1}, {0, 0, 0, 0, 0}
              },
              new int[] {0, 4},
              new int[] {3, 2},
              false),
          arguments(
              new int[][] {
                {0, 0, 0, 0, 0}, {1, 1, 0, 0, 1}, {0, 0, 0, 0, 0}, {0, 1, 0, 0, 1}, {0, 1, 0, 0, 0}
              },
              new int[] {4, 3},
              new int[] {0, 1},
              false));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void hasPath(int[][] input, int[] start, int[] end, boolean expectedResult) {

    boolean actualResult = solution.hasPath(input, start, end);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
